Refactor code structure: created socket_handler and separated sockets from route files, moved socketio import to context, removed clients.py and client.py, updated task_manager and wav_task classes
