+++ app.py
@@ -1,5 +1,5 @@
 from app import create_app
-from app.extensions import socketio
+from app.context import socketio
 
 app = create_app()
