+++ app/__init__.py
@@ -1,10 +1,13 @@
 from flask import Flask
 from flask_cors import CORS
 from app.config import Config
-from app.extensions import socketio
+from app.context import socketio
+from app.socket_handler import *
 from app.blueprints.home import home_bp
 from app.blueprints.wav_to_text import wav_bp
 from app.blueprints.voice_to_text import voice_bp
+from app.data.task_manager import TaskManager
+
 import logging
 
 logging.basicConfig(level=logging.INFO)
@@ -16,6 +19,9 @@ def create_app() -> Flask:
     try:
         app = Flask(__name__)
         app.config.from_object(Config)
+        
+        # Инициализация TaskManager и сохранение его в контексте приложения
+        app.config['TASK_MANAGER'] = TaskManager()
         
         # Initialize SocketIO
         socketio.init_app(app)
