+++ app/blueprints/home.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request, jsonify, render_template
-from app.extensions import socketio
+from app.context import socketio
 from app.config import Config
 
 home_bp = Blueprint('home_bp', __name__, template_folder="templates")
