+++ app/tempo/client.py
@@ -2,5 +2,3 @@ class Client:
     def __init__(self, sid: str, ip: str):
         self.sid = sid
         self.ip = ip
-
-
